+++ app/src/main/java/com/chaddy50/musicapp/data/entity/Genre.kt
@@ -19,7 +19,7 @@ import androidx.room.PrimaryKey
 )
 data class Genre(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val name: String,
-    val parentGenreId: Int? = null
+    val parentGenreId: Long? = null
 )
